requirements.txt
anthropic
requests
python-dotenv
schedule
flask
feedparser
Pillow
moviepy
imageio
imageio-ffmpeg
